Move ApplicationStatusInfo and StatusOverview back into _private
